feat: components live at /components/<slug>, /preview stays the fixture
The canonical page for each of 228 components moves to /components/<slug>.
"Preview" reads as a temporary demo, and it was the word in every URL the
launch push would earn a backlink to. Doing this now is nearly free precisely
because nothing links here yet: the site is not indexed and has no external
backlinks. After the launch it would mean redirects forever.

/preview/<slug> does not redirect and does not go away. It stays as the
verification and recording fixture, now noindex with a canonical pointing at
the real page. Two attempts to remove it both failed on measurement rather
than on taste:

- Pointing the gate at /components/<slug> breaks it, because that page now
  correctly has site chrome, and verify.ts takes the first visible
  "button, a, [role=button]" as the component's own control. Measured: it
  grabs the sidebar's ns-ui wordmark. Every hover, press, focus and
  gate.openBy screenshot would have been driven against navigation.
- Pointing it at /preview/<slug>/embed breaks it differently: that route is
  always inert and always autoplaying, by design. Measured with Playwright,
  focus never leaves document.body there, so all 228 components with a
  control would false-fail "nothing is reachable by Tab", and the
  hover/focus screenshot diffs would be contaminated by animation the test
  did not cause.

So the gate does not change at all. scripts/verify.ts, scripts/record.ts and
next.config.ts are byte-identical to before this commit - a deliberate result,
not a coincidence. Destabilising the thing that catches real defects across
228 components, as a side effect of a rename, is a bad trade.

demo-frame.tsx holds the one implementation both routes render, so the demo
logic cannot drift between them, while metadata and JSON-LD stay per-route
because those are exactly what should differ. /embed keeps its own five-line
implementation: its docblock already rejected shared-component indirection
there, and that call still holds.

/r/<slug>.json is untouched and always will be. It is what npx shadcn add
fetches, it is baked into two published npm packages and llms.txt, and every
install command anyone has already run points at it.

Measured: /preview/<slug> 200 bare with noindex + canonical and zero JSON-LD,
/components/<slug> 200 with chrome and two JSON-LD blocks, /embed 200 with no
redirect, /r/<slug>.json unchanged, sitemap 234 URLs, C6 197,248 bytes against
a 206,509 ceiling. Full 228-component gate: 227 pass. The one failure,
undo-ghost-row, reproduces identically at HEAD before this change and is not
caused by it.
