+++ app/preview/[name]/embed/page.tsx
@@ -3,28 +3,34 @@ import { notFound } from "next/navigation";
 import { demos } from "@/registry/index";
 import autoplayMap from "@/lib/autoplay.generated.json";
 import { parseAutoplay } from "@/lib/autoplay";
-import { AutoplayDriver } from "../autoplay-driver";
+import { AutoplayDriver } from "@/app/_components/autoplay-driver";
 
 /**
  * The card thumbnail route — what every catalog and featured card loads in its
  * iframe.
  *
- * It exists only to be *cacheable*. The sibling `/preview/<name>` renders the
- * identical thing for `?embed=1&autoplay=1`, but reading `searchParams` makes
- * that route fully dynamic: it never enters the prerender manifest, so every
- * card frame was a function invocation serving `no-store` (measured:
+ * It exists only to be *cacheable*. `/preview/<name>` — the verification/
+ * recording fixture, see that route's own docblock — renders the identical
+ * thing for `?embed=1&autoplay=1`, but reading `searchParams` makes that
+ * route fully dynamic: it never enters the prerender manifest, so every card
+ * frame was a function invocation serving `no-store` (measured:
  * `x-vercel-cache: MISS` on every one, ~4-12 of them per homepage view). Same
  * markup with the two flags baked into the path instead of the query string
  * prerenders and serves from the CDN.
  *
  * `/preview/<name>` is deliberately left untouched — `scripts/verify.ts` and
- * `scripts/record.ts` screenshot it directly, and `?embed=1&interactive=1`
- * (the playground frame) still needs the dynamic variant.
+ * `scripts/record.ts` screenshot it directly (not this route: `inert` and the
+ * unconditional autoplay below would break the gate's Tab-reachability and
+ * hover/focus screenshot diffs — see `/preview/<name>`'s docblock for the
+ * measured reasons), and `?embed=1&interactive=1` (the playground frame)
+ * still needs the dynamic variant.
  *
- * Kept in lockstep with the sibling by hand: it is five lines of JSX, and the
- * shared-component indirection costs more than it saves. The invariant is that
- * a card and the reference page render the same DOM — if that drifts, the
- * screenshot gate catches it.
+ * Kept in lockstep with the fixture by hand rather than sharing `DemoFrame`:
+ * it is five lines of JSX, always `inert` and always autoplaying regardless
+ * of query params (unlike `DemoFrame`'s conditional logic), and the
+ * shared-component indirection costs more than it saves. The invariant is
+ * that a card and the reference fixture render the same DOM for
+ * `?embed=1&autoplay=1` — if that drifts, the screenshot gate catches it.
  */
 export const revalidate = 3600;
