+++ app/preview/[name]/page.tsx
@@ -1,28 +1,41 @@
 import type { Metadata } from "next";
-import { notFound } from "next/navigation";
-import { demos } from "@/registry/index";
 import registry from "@/registry.json";
-import autoplayMap from "@/lib/autoplay.generated.json";
-import { parseAutoplay } from "@/lib/autoplay";
-import { AutoplayDriver } from "./autoplay-driver";
 import { REGISTRY_ORIGIN } from "@/lib/registry-origin";
-import { jsonLdScript } from "@/lib/json-ld";
-import pkg from "@/package.json";
+import { DemoFrame } from "@/app/_components/demo-frame";
 
-// package.json's repository.url is the git clone URL (`...ns-ui.git`);
-// schema.org's `codeRepository` wants the browsable repo URL instead.
-const CODE_REPOSITORY = pkg.repository.url.replace(/\.git$/, "");
-const LICENSE_URL =
-  pkg.license === "MIT"
-    ? "https://opensource.org/licenses/MIT"
-    : undefined;
-
-// The per-component opengraph-image.tsx in this same folder is picked up by
-// the file-convention automatically — this only needs to supply the title
-// and description text, both openGraph and twitter fall back to `title`/
-// `description` unless overridden, so they're set explicitly instead to
-// guarantee og:title/twitter:title are correct rather than relying on that
-// implicit resolution.
+/**
+ * The verification/recording fixture. Chrome-less (matched by
+ * `isBarePreview` in `app/_components/site-shell.tsx`), noindex, and
+ * canonical back to `/components/<name>` — deliberately not the page a
+ * visitor or a crawler is meant to land on. `scripts/verify.ts` and
+ * `scripts/record.ts` navigate here directly, and nowhere else in the app
+ * links to it.
+ *
+ * `/components/<name>` (the canonical, chrome-full, indexed page — see that
+ * route's own docblock) shares this exact rendering via `DemoFrame`; this
+ * file differs from it only in metadata (noindex + canonical, no JSON-LD)
+ * and in staying outside site chrome.
+ *
+ * Two measured reasons neither of the other candidate gate targets works —
+ * recorded here so the next person doesn't redo the experiment:
+ *
+ *  - Pointing the gate at `/components/<name>` (chrome-full) breaks the
+ *    "first visible interactive element" locator that drives the hover/
+ *    press/focus/`gate.openBy` screenshots in `verify.ts`: measured, it
+ *    resolves to the sidebar's own wordmark link, not anything belonging to
+ *    the component.
+ *  - Pointing it at `/preview/<name>/embed` breaks Tab-reachability:
+ *    `/embed` is always `inert` (see `demo-frame.tsx`'s embed handling) and
+ *    always runs autoplay unconditionally, so `verify.ts`'s "Tab up to 12
+ *    times and land on something" check never lands (focus measured staying
+ *    on `document.body`), and any interaction screenshot would be
+ *    contaminated by motion the driver is already running on its own.
+ *
+ * So this route stays a plain, uninert, non-autoplaying render — the only
+ * shape that satisfies what the gate actually asserts — while `/components/
+ * <name>` carries the chrome and the structured data that make the page
+ * worth indexing.
+ */
 export async function generateMetadata({
   params,
 }: {
@@ -32,18 +45,16 @@ export async function generateMetadata({
   const item = registry.items.find((i) => i.name === name);
   if (!item) return {};
 
-  const title = `${item.title} — ns-ui`;
-  const description = item.description;
-
+  const canonical = `${REGISTRY_ORIGIN}/components/${name}`;
   return {
-    title,
-    description,
-    openGraph: { title, description, type: "website" },
-    twitter: { card: "summary_large_image", title, description },
+    title: `${item.title} — ns-ui`,
+    description: item.description,
+    robots: { index: false, follow: false },
+    alternates: { canonical },
   };
 }
 
-export default async function PreviewPage({
+export default async function PreviewFixturePage({
   params,
   searchParams,
 }: {
@@ -52,93 +63,5 @@ export default async function PreviewPage({
 }) {
   const { name } = await params;
   const { embed, autoplay, interactive } = await searchParams;
-  const Demo = demos[name];
-  if (!Demo) notFound();
-
-  const item = registry.items.find((i) => i.name === name);
-  const pageUrl = `${REGISTRY_ORIGIN}/preview/${name}`;
-
-  // Only fields populated from real registry/package data — nothing here is
-  // invented to fill out the schema. Server-rendered (this is an async
-  // Server Component), so crawlers see it in the initial HTML.
-  const softwareSourceCodeJsonLd = item
-    ? {
-        "@context": "https://schema.org",
-        "@type": "SoftwareSourceCode",
-        name: item.title,
-        description: item.description,
-        url: pageUrl,
-        programmingLanguage: "TypeScript",
-        codeRepository: CODE_REPOSITORY,
-        ...(LICENSE_URL ? { license: LICENSE_URL } : {}),
-      }
-    : null;
-
-  const breadcrumbJsonLd = item
-    ? {
-        "@context": "https://schema.org",
-        "@type": "BreadcrumbList",
-        itemListElement: [
-          { "@type": "ListItem", position: 1, name: "ns-ui", item: REGISTRY_ORIGIN },
-          { "@type": "ListItem", position: 2, name: item.title, item: pageUrl },
-        ],
-      }
-    : null;
-
-  // `?embed=1` is how the landing-page cards load this page inside an iframe.
-  // It changes nothing visual — this page stays the reference the cards are
-  // matched against — it only makes the demo inert. Without it, a demo that
-  // focuses something on mount (command-palette-orbit focuses its input) hands
-  // focus to the iframe, and the browser scrolls the *host* page to reveal
-  // that iframe: the landing page jumped ~1000px on its own. Inert also keeps
-  // the demo's own controls out of the host page's tab order.
-  const embedded = embed === "1";
-
-  // `&interactive=1` is the featured-card "Interact" gesture: the visitor
-  // already clicked to opt in, so the mount-time-focus hazard above cannot
-  // recur (the frame is already on screen and focus does not have to scroll
-  // anything into view). Only takes effect inside an embed; a bare
-  // `/preview/<name>?interactive=1` behaves exactly like the honest
-  // reference page.
-  const interactiveEmbed = embedded && interactive === "1";
-
-  // `&autoplay=1` additionally runs the shared driver (see ./autoplay-driver):
-  // it synthesises the input a component needs so a card demonstrates itself
-  // instead of freezing on a still frame. Embed-only and descriptor-only —
-  // without both params, or without an `autoplay` key in that component's
-  // meta.json, nothing mounts and this page behaves exactly as it always has.
-  // `inert` is unchanged in autoplay mode: the driver dispatches events
-  // directly to target elements, which `inert` does not block. Interactive
-  // mode never autoplays — the visitor is driving directly.
-  // No cast: this is JSON read off disk, so its literal types are widened
-  // (`number[]`, not `[number, number]`). Asserting AutoplayMap onto it only
-  // lies about a shape TS can't confirm — parseAutoplay already validates at
-  // runtime and returns null for anything malformed.
-  const spec =
-    embedded && autoplay === "1" && !interactiveEmbed
-      ? parseAutoplay((autoplayMap as Record<string, unknown>)[name])
-      : null;
-
-  return (
-    <div
-      className="min-h-screen"
-      inert={embedded && !interactiveEmbed}
-      data-autoplay-root={spec ? "" : undefined}
-    >
-      {softwareSourceCodeJsonLd ? (
-        <script
-          type="application/ld+json"
-          dangerouslySetInnerHTML={{ __html: jsonLdScript(softwareSourceCodeJsonLd) }}
-        />
-      ) : null}
-      {breadcrumbJsonLd ? (
-        <script
-          type="application/ld+json"
-          dangerouslySetInnerHTML={{ __html: jsonLdScript(breadcrumbJsonLd) }}
-        />
-      ) : null}
-      <Demo />
-      {spec ? <AutoplayDriver spec={spec} /> : null}
-    </div>
-  );
+  return <DemoFrame name={name} embed={embed} autoplay={autoplay} interactive={interactive} />;
 }
